use double as preferred simd sorting type on x86

+++ src/lib/operators/join_simd_sort_merge.hpp
@@ -45,9 +45,9 @@ class JoinSimdSortMerge : public AbstractJoinOperator {
  protected:
 // Datatype used for simd sorting (has to be 64 bits).
 #if defined(__AVX512F__)
-  using SortingType = int64_t;
+  using SortingType = double;
 #elif defined(__AVX2__)
-  using SortingType = int64_t;
+  using SortingType = double;
 #elif defined(__powerpc__) || defined(__ppc__) || defined(_ARCH_PPC)
   using SortingType = int64_t;
 #elif defined(__arm__) || defined(__aarch64__)
